feat: add cashback update api and admin operation support

+++ app/models/cashback_record.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, Text
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Text, Numeric
 from sqlalchemy.sql import func
 
 from app.core.db import Base
@@ -12,8 +12,8 @@ class CashbackRecord(Base):
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
     order_id = Column(Integer, ForeignKey("orders.id"), nullable=False, index=True)
 
-    expected_cashback_amount = Column(Float, default=0.0)
-    actual_cashback_amount = Column(Float, default=0.0)
+    expected_cashback_amount = Column(Numeric(12, 2), default=0.00)
+    actual_cashback_amount = Column(Numeric(12, 2), default=0.00)
 
     status = Column(String(50), default="pending")
     remark = Column(Text, nullable=True)
